feat: 결제기능 구현
- 결제 기능 구현

+++ product-service/src/main/java/com/hanghae/productservice/dto/StockProductDto.java
@@ -8,7 +8,7 @@ import lombok.Getter;
 @AllArgsConstructor
 public class StockProductDto {
     private long productId;
-    private int stock;
+    private long stock;
 
     public StockProductDto(Product product) {
         this.productId = product.getId();
